add tfoot tag
The <tfoot> tag can appear either above or below <tbody>, so need to
permit use of rows from the top or bottom of the rows array.  This
adds extra conditionals to the loop over rows.

The check at the end of the loop over rows to add the closing </thead>
has been moved into the check at the top of the loop during the next
loop iteration.