bugfix: cells which tested boolean false (e.g. "0") were not output
When concatenating the strings making up the table,

     join( "",  grep { $_ } @table )

strings which didn't pass the Perl bool truth test
were excluded, which included legitimate strings composed
of the string "0".

The original code was perhaps designed to skip undefined strings,
but that can no longer happen.  Removing the test fixes the
bug and makes things a bit faster.